+++ src/main/java/org/opengis/cite/geomatics/time/TemporalUtils.java
@@ -1,20 +1,24 @@
 package org.opengis.cite.geomatics.time;
 
+import java.util.TreeSet;
+
+import org.geotoolkit.temporal.factory.DefaultTemporalFactory;
 import org.opengis.temporal.Instant;
 import org.opengis.temporal.Period;
 import org.opengis.temporal.RelativePosition;
+import org.opengis.temporal.TemporalFactory;
 import org.opengis.temporal.TemporalGeometricPrimitive;
 
 /**
- * Provides specialized assertion methods that apply to representations of
- * temporal objects.
+ * Provides various utility methods that apply to representations of temporal
+ * objects.
  * 
  * @see "ISO 19108: Geographic information -- Temporal schema"
  * @see <a
  *      href="http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.329.2647">Maintaining
  *      knowledge about temporal intervals</a>
  */
-public class TemporalAssert {
+public class TemporalUtils {
 
 	/**
 	 * Asserts that the first temporal primitive is related to the second one so
@@ -66,7 +70,7 @@ public class TemporalAssert {
 	 *            A temporal geometric primitive (instant or period).
 	 * @return A String that displays temporal position in accord with ISO 8601.
 	 */
-	static String temporalGeometricPrimitiveToString(
+	public static String temporalGeometricPrimitiveToString(
 			TemporalGeometricPrimitive timeObj) {
 		StringBuilder strBuilder = new StringBuilder();
 		if (Instant.class.isInstance(timeObj)) {
@@ -79,9 +83,44 @@ public class TemporalAssert {
 			strBuilder
 					.append(period.getBeginning().getPosition().getDateTime())
 					.append('/');
-			strBuilder
-					.append(period.getBeginning().getPosition().getDateTime());
+			strBuilder.append(period.getEnding().getPosition().getDateTime());
 		}
 		return strBuilder.toString();
+	}
+
+	/**
+	 * Determines the total temporal extent of a set of temporal primitives.
+	 * 
+	 * @param tmSet
+	 *            An ordered set of TemporalGeometricPrimitive objects (instant
+	 *            or period).
+	 * @return A period that spans the entire time range of the set members.
+	 */
+	public static Period temporalExtent(
+			TreeSet<TemporalGeometricPrimitive> tmSet) {
+		TemporalGeometricPrimitive first = tmSet.first();
+		TemporalGeometricPrimitive last = tmSet.last();
+		Instant startOfPeriod;
+		if (first instanceof Instant) {
+			startOfPeriod = Instant.class.cast(first);
+		} else {
+			startOfPeriod = Period.class.cast(first).getBeginning();
+		}
+		Instant endOfPeriod;
+		if (last instanceof Instant) {
+			endOfPeriod = Instant.class.cast(last);
+			// check if last DURING first
+			if ((first instanceof Period)
+					&& endOfPeriod.relativePosition(
+							Period.class.cast(first).getEnding()).equals(
+							RelativePosition.BEFORE)) {
+				endOfPeriod = Period.class.cast(first).getEnding();
+			}
+		} else {
+			endOfPeriod = Period.class.cast(last).getEnding();
+		}
+		TemporalFactory tmFactory = new DefaultTemporalFactory();
+		Period period = tmFactory.createPeriod(startOfPeriod, endOfPeriod);
+		return period;
 	}
 }